ТЗ 12(3 коммит) Добавления базы данных

+++ src/main/java/ru/yandex/practicum/filmorate/dal/mappers/FilmRowMapper.java
@@ -21,7 +21,7 @@ public class FilmRowMapper implements RowMapper<Film> {
     @Override
     public Film mapRow(ResultSet rs, int rowNum) throws SQLException {
         Film film = new Film();
-        film.setId(rs.getLong("film_id"));
+        film.setId(rs.getInt("film_id"));
         film.setName(rs.getString("name"));
         film.setDescription(rs.getString("description"));
         film.setReleaseDate(rs.getDate("release_date").toLocalDate());
